All magesteps use same go.mod; add docs.
