+++ .github/workflows/publish-sdk-nodejs.yml
@@ -10,6 +10,6 @@ jobs:
       - uses: actions/setup-go@v3
         with:
           go-version: 1.19
-      - run: ./hack/make sdk:nodejs:publish ${{ github.ref_name }}
+      - run: ./hack/make-prod sdk:nodejs:publish ${{ github.ref_name }}
         env:
           NPM_TOKEN: ${{ secrets.RELEASE_NPM_TOKEN }}
