--- hack/bootstrap
@@ -1,24 +1,0 @@
-#!/usr/bin/env bash
-
-set -e -u -x
-
-ROOTDIR="$(cd $(dirname "${BASH_SOURCE[0]}")/.. && pwd)"
-HACK="$ROOTDIR/hack"
-MAGEFILES="$ROOTDIR/magefiles"
-
-function cleanup() {
-  rm -f $MAGEFILES/go.{mod,sum}
-}
-
-ln -sf $MAGEFILES/mod/go.mod $MAGEFILES/go.mod
-ln -sf $MAGEFILES/mod/go.sum $MAGEFILES/go.sum
-
-trap cleanup EXIT
-
-pushd $MAGEFILES
-  go run $MAGEFILES/main.go -compile /tmp/dagger-mage-bootstrap
-popd
-
-cleanup
-
-/tmp/dagger-mage-bootstrap "$@"
