+++ internal/mage/docs.go
@@ -1,10 +1,10 @@
-package targets
+package mage
 
 import (
 	"context"
 	"os"
 
-	"github.com/dagger/dagger/magefiles/util"
+	"github.com/dagger/dagger/internal/mage/util"
 
 	"dagger.io/dagger"
 	"github.com/magefile/mage/mg"
