+++ internal/mage/main.go
