+++ internal/mage/sdk/go.go
@@ -9,7 +9,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/dagger/dagger/magefiles/util"
+	"github.com/dagger/dagger/internal/mage/util"
 
 	"dagger.io/dagger"
 	"github.com/magefile/mage/mg"
