--- magefiles/mod/go
@@ -1,18 +1,0 @@
-#!/usr/bin/env bash
-
-set -e -u -x
-
-MAGEFILES=$(cd $(dirname $0)/.. && pwd)
-
-function cleanup() {
-  rm -f $MAGEFILES/go.{mod,sum}
-}
-
-ln -sf $MAGEFILES/mod/go.mod $MAGEFILES/go.mod
-ln -sf $MAGEFILES/mod/go.sum $MAGEFILES/go.sum
-
-trap cleanup EXIT
-
-cd $MAGEFILES
-
-go "$@"
